refactor returnOnInvestmentListItems
